pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-tenant-authx"
version = "0.1.0"
description = "Tenant-aware authentication and authorization for Django multi-tenant SaaS applications"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "RAJID K K", email = "rajidkk34@gmail.com"}
]
maintainers = [
    {name = "RAJID K K", email = "rajidkk34@gmail.com"}
]
keywords = [
    "django",
    "multi-tenant",
    "authentication",
    "authorization",
    "saas",
    "rbac",
    "permissions",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.2",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"
dependencies = [
    "Django>=4.2",
]

[project.optional-dependencies]
drf = [
    "djangorestframework>=3.14",
]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "hypothesis>=6.0",
    "coverage>=7.0",
    "flake8>=6.0",
    "black>=23.0",
    "mypy>=1.0",
    "django-stubs>=4.0",
    "djangorestframework-stubs>=3.0",
]
all = [
    "django-tenant-authx[drf,dev]",
]

[project.urls]
Homepage = "https://github.com/kkrajid/django-tenant-authx"
Documentation = "https://github.com/kkrajid/django-tenant-authx#readme"
Repository = "https://github.com/kkrajid/django-tenant-authx.git"
Changelog = "https://github.com/kkrajid/django-tenant-authx/blob/main/CHANGELOG.md"
Issues = "https://github.com/kkrajid/django-tenant-authx/issues"

[tool.setuptools]
packages = ["tenant_authx", "tenant_authx.drf", "tenant_authx.migrations"]
include-package-data = true

[tool.setuptools.package-data]
tenant_authx = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312", "py313"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.venv
    | build
    | dist
    | migrations
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]
exclude = ["migrations", "tests"]

[[tool.mypy.overrides]]
module = "tenant_authx.*"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "tests.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.coverage.run]
branch = true
source = ["tenant_authx"]
omit = ["*/migrations/*", "*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
